+++ R/provider-openai.R
@@ -295,7 +295,7 @@ method(as_json, list(ProviderOpenAI, Turn)) <- function(provider, x) {
     c(user, tools)
   } else if (x@role == "assistant") {
     # Tool requests come out of content and go into own argument
-    is_tool <- map_lgl(x@contents, S7_inherits, ContentToolRequest)
+    is_tool <- map_lgl(x@contents, is_tool_request)
     content <- as_json(provider, x@contents[!is_tool])
     tool_calls <- as_json(provider, x@contents[is_tool])